Everything (Essential) is finish. Cheers, mate.
Although some part are scrapped. I hoped I could return and make it
better after all this has ended.

Would be nice to be able to choose directory
via split '/' and check if it's exist (though it's a long shot.)and
separate text files with randomized word with its module and add it at
the start of the text.